Server/requirements.txt
blinker==1.6.2
click==8.1.3
dnspython==2.3.0
Flask==2.3.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
pymongo==4.3.3
python-dotenv==1.0.0
Werkzeug==2.3.4
